+++ build.clj
@@ -13,7 +13,7 @@
 (def lib 'org.scicloj/metamorph.ml)
 ; alternatively, use MAJOR.MINOR.COMMITS:
 ;; (def version (format "6.2.%s" (b/git-count-revs nil)))
-(def version "0.10.4")
+(def version "0.12")
 (def class-dir "target/classes")
 (def basis (b/create-basis {:project "deps.edn"}))
 (def jar-file (format "target/%s-%s.jar" (name lib) version))
@@ -40,7 +40,7 @@
     [:connection "scm:git:https://github.com/scicloj/metamorph.ml.git"]
     [:developerConnection "scm:git:https://github.com/scicloj/metamorph.ml.git"]
 
-    [:tag (str "v" version)]]])
+    [:tag (str version)]]])
 
 
 (defn jar [_]